lp_python/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv requirements
#

-i https://pypi.org/simple
astroid==2.11.7; python_full_version >= '3.6.2'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cachetools==5.2.0; python_version ~= '3.7'
certifi==2022.6.15; python_version >= '3.6'
cfgv==3.3.1; python_full_version >= '3.6.1'
charset-normalizer==2.1.0; python_version >= '3.6'
dill==0.3.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
distlib==0.3.5
filelock==3.7.1; python_version >= '3.7'
google-api-core==2.8.2; python_version >= '3.6'
google-auth==2.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-core==2.3.2; python_version >= '3.7'
google-cloud-storage==2.4.0
google-cloud-vision==3.0.0
google-crc32c==1.3.0; python_version >= '3.6'
google-resumable-media==2.3.3; python_version >= '3.6'
googleapis-common-protos==1.56.4; python_version >= '3.7'
grpcio-status==1.47.0
grpcio==1.47.0
identify==2.5.2; python_version >= '3.7'
idna==3.3; python_version >= '3.5'
iniconfig==1.1.1
isort==5.10.1; python_version < '4.0' and python_full_version >= '3.6.1'
lazy-object-proxy==1.7.1; python_version >= '3.6'
mccabe==0.7.0; python_version >= '3.6'
mysql-connector-python==8.0.29
nodeenv==1.7.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
numpy==1.23.1; python_version < '3.10' and platform_machine != 'aarch64' and platform_machine != 'arm64'
packaging==21.3; python_version >= '3.6'
pandas==1.4.3
platformdirs==2.5.2; python_version >= '3.7'
pluggy==1.0.0; python_version >= '3.6'
pre-commit==2.20.0
proto-plus==1.20.6; python_version >= '3.6'
protobuf==3.20.1; python_version >= '3.7'
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pylint==2.14.5
pyparsing==3.0.9; python_full_version >= '3.6.8'
pytest==7.1.2
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.20.0
pytz==2022.1
pyyaml==6.0; python_version >= '3.6'
requests==2.28.1
rsa==4.9; python_version >= '3.6'
setuptools==63.2.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1; python_version >= '3.7'
tomlkit==0.11.1; python_version >= '3.6' and python_version < '4.0'
typing-extensions==4.3.0; python_version < '3.10'
urllib3==1.26.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4.0'
virtualenv==20.15.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wrapt==1.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
